feat(common): Moved getImages function from project.js to common.js to be used in other modules
