+++ lib/common.js
@@ -1,11 +1,16 @@
+const { Buffer } = require('buffer');
 const path = require('path');
 const queryString = require('querystring');
 const url = require('url');
+const fs = require('fs');
+const https = require('https');
 const { marked } = require('marked');
 const { JSDOM } = require('jsdom');
+const sharp = require('sharp');
 
 
 const internals = {};
+const { stat, readFile, writeFile } = fs.promises;
 
 
 internals.parseMeta = (metaToken, metaKeys, links) => {
@@ -252,6 +257,43 @@ const parseLink = (document, el) => {
 };
 
 
+const getCoverImage = (tokens) => {
+  const firstTokenWithImg = tokens.find(token => (
+    token.type === 'paragraph'
+    && (
+      /!\[.+\]\(.+\)/g.test(token.text)
+      || /<img /g.test(token.text)
+    )
+  ));
+
+  if (!firstTokenWithImg) {
+    return null;
+  }
+
+  const html = marked.parse(firstTokenWithImg.text);
+  const { window } = new JSDOM(html);
+  return window.document.querySelector('img').src;
+};
+
+
+const fetchImage = src => new Promise((resolve, reject) => {
+  https.get(src, (resp) => {
+    if (resp.statusCode !== 200) {
+      return reject(new Error(`HTTP error ${resp.statusCode}`));
+    }
+
+    let buffer;
+    resp.on('error', reject);
+    resp.on('data', (chunk) => {
+      buffer = (typeof buffer === 'undefined')
+        ? chunk
+        : Buffer.concat([buffer, chunk]);
+    });
+    return resp.on('end', () => resolve(buffer));
+  });
+});
+
+
 internals.parseLinks = (body) => {
   const { window } = new JSDOM(body);
   const { document } = window;
@@ -402,3 +444,33 @@ exports.getReadmeFileName = locale => (
     ? 'README.md'
     : `README.${locale}.md`
 );
+
+exports.getImages = (dir, tokens) => {
+  const cover = getCoverImage(tokens);
+  const thumbFile = path.join(dir, 'thumb.png');
+
+  const createThumb = async (src) => {
+    const buffer = await fetchImage(src);
+    const data = await sharp(buffer).resize(395).toBuffer();
+    await writeFile(thumbFile, data, { encoding: 'base64' });
+    return data;
+  };
+
+  return stat(thumbFile)
+    .then(() => readFile(thumbFile, 'base64'))
+    .catch((err) => {
+      if (err.code !== 'ENOENT') {
+        throw err;
+      }
+      // does not exist!
+      return !cover ? null : createThumb(cover);
+    })
+    .then(data => ({
+      cover,
+      thumb: (
+        data
+          ? `data:image/png;base64,${data.toString('base64')}`
+          : null
+      ),
+    }));
+};
