+++ lib/project.js
@@ -1,16 +1,12 @@
-const { Buffer } = require('buffer');
 const path = require('path');
 const fs = require('fs');
-const https = require('https');
 const { marked } = require('marked');
 const yaml = require('js-yaml');
-const { JSDOM } = require('jsdom');
-const sharp = require('sharp');
-const { getReadmeFileName } = require('./common');
+const { getReadmeFileName, getImages } = require('./common');
 const pkg = require('../package.json');
 
 
-const { stat, readFile, writeFile } = fs.promises;
+const { readFile } = fs.promises;
 
 
 const getProjectTitle = ([token]) => {
@@ -79,6 +75,7 @@ const findChildren = (tree, nodePath) => {
   return findChildren(found, parts.slice(1).join('/'));
 };
 
+
 const findFlattenedChildren = (known, learningObjective) => {
   const children = findChildren(known, learningObjective);
   if (!children) {
@@ -146,74 +143,6 @@ const getSummary = (tokens) => {
   }
 
   return marked.parse(tokens[headingIdx + 1].text);
-};
-
-
-const fetchImage = src => new Promise((resolve, reject) => {
-  https.get(src, (resp) => {
-    if (resp.statusCode !== 200) {
-      return reject(new Error(`HTTP error ${resp.statusCode}`));
-    }
-
-    let buffer;
-    resp.on('error', reject);
-    resp.on('data', (chunk) => {
-      buffer = (typeof buffer === 'undefined')
-        ? chunk
-        : Buffer.concat([buffer, chunk]);
-    });
-    return resp.on('end', () => resolve(buffer));
-  });
-});
-
-
-const getCoverImage = (tokens) => {
-  const firstTokenWithImg = tokens.find(token => (
-    token.type === 'paragraph'
-    && (
-      /!\[.+\]\(.+\)/g.test(token.text)
-      || /<img /g.test(token.text)
-    )
-  ));
-
-  if (!firstTokenWithImg) {
-    return null;
-  }
-
-  const html = marked.parse(firstTokenWithImg.text);
-  const { window } = new JSDOM(html);
-  return window.document.querySelector('img').src;
-};
-
-
-const getImages = (dir, tokens) => {
-  const cover = getCoverImage(tokens);
-  const thumbFile = path.join(dir, 'thumb.png');
-
-  const createThumb = async (src) => {
-    const buffer = await fetchImage(src);
-    const data = await sharp(buffer).resize(395).toBuffer();
-    await writeFile(thumbFile, data, { encoding: 'base64' });
-    return data;
-  };
-
-  return stat(thumbFile)
-    .then(() => readFile(thumbFile, 'base64'))
-    .catch((err) => {
-      if (err.code !== 'ENOENT') {
-        throw err;
-      }
-      // does not exist!
-      return !cover ? null : createThumb(cover);
-    })
-    .then(data => ({
-      cover,
-      thumb: (
-        data
-          ? `data:image/png;base64,${data.toString('base64')}`
-          : null
-      ),
-    }));
 };
